chore: renamed the cacheTest file
